requirements.txt
pyyaml
pytest
pytest-mock